+++ src/Network/Network/ChangeLog.md
@@ -48,6 +48,9 @@
     - `Update-AzVirtualHub`
 * Updated cmdlets to expose two read-only properties of client certificate.
     - `Get-AzApplicationGatewayTrustedClientCertificate`
+* Updated below cmdlets to enable or disable BgpRouteTranslationForNat on VpnGateway.
+    - `New-AzVpnGateway`
+	- `Update-AzVpnGateway`
 
 ## Version 4.8.0
 * Updated validation to allow passing zero value for saDataSizeKilobytes parameter